+++ CHANGELOG.md
@@ -26,7 +26,8 @@ Classify the change according to the following categories:
 ## Develop 2024-07-6
 ### Added
 - Added new inputs **om_cost_per_kw** and **om_cost_per_kwh** to `ElectricStorage` for modeling capacity-based O&M 
-- Added new input **per_timestep_self_discharge_fraction** to `ElectricStorage` for modeling battery self-discharge
+- Added new input **soc_based_per_ts_self_discharge_fraction** to `ElectricStorage` for modeling SOC-based battery self-discharge
+- Added new input **capacity_based_per_ts_self_discharge_fraction** to `ElectricStorage` for modeling capacity-based battery self-discharge
 - Added new outputs **lifecycle_om_cost_after_tax** and **year_one_om_cost_before_tax** to `ElectricStorage` 
 - Added testsets **Electric Storage O&M** and **Electric Storage Self-Discharge** in `test/runtests.jl`